+++ CHANGELOG.md
@@ -124,6 +124,7 @@
 - Tracking Number : change in stock move and stock move line based on tracking number configuration.
 - StockMove : Include sub stock location's lines in Report. 
 - Notification : Improve notification management. 
+- Manage origin date on supplier invoice. This date is used for compute due date and is used in account move line to fill the origin date.
 
 ## Improvements
 - Move project resource menu entry to project configuration menu.